Removed SectionTitle component. Fixed locked/optional logic for lock/delete buttons. Edited the description section to use the changed section components/structure.
