+++ client/src/pages/board/components/dashboardForms/SectionContent.js
@@ -2,7 +2,6 @@ import React from 'react';
 import WithIcon from '../../../../components/WithIcon';
 
 const SectionContent = ({locked, children}) => {
-  console.log(locked);
   return <WithIcon aligner>{React.cloneElement(children, {locked})}</WithIcon>;
 };
